% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box-colors.R
\name{box_docs_color}
\alias{box_docs_color}
\title{Get color for documentation infobox}
\usage{
box_docs_color(num_docs)
}
\arguments{
\item{num_docs}{The number of documents.}
}
\description{
Gets the color for the documentation
infobox based on number of documents.
}
