% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box-colors.R
\name{box_meta_color}
\alias{box_meta_color}
\title{Get color for metadata infobox}
\usage{
box_meta_color(num_meta_files)
}
\arguments{
\item{num_meta_files}{The number of metadata files.}
}
\description{
Gets the color for the metadata infobox
based on the number of metadata files.
}
