% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box-colors.R
\name{box_success_color}
\alias{box_success_color}
\title{Get color for success rate infobox}
\usage{
box_success_color(success_rate)
}
\arguments{
\item{success_rate}{Integer percentage from 0 to 100.}
}
\description{
Gets the color for the success rate infobox
based on the percent success rate.
}
