% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine-all-metadata.R
\name{combine_all_metadata}
\alias{combine_all_metadata}
\title{Combine all metadata}
\usage{
combine_all_metadata(fileview)
}
\arguments{
\item{fileview}{The fileview for a specific study.}
}
\description{
Combine all metadata into a single dataframe. Note that this function does
not handle errors when trying to combine data. For example, if there are
no common columns between the metadata files, then the function will throw
an error.
}
