% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file-info-table.R
\name{create_info_table}
\alias{create_info_table}
\title{Create a nicer version of the file info}
\usage{
create_info_table(fileview, syn)
}
\arguments{
\item{fileview}{The fileview table.}

\item{syn}{Synapse client object.}
}
\description{
Updates come default Synapse columns to be more useful.
The column 'currentVersion' is changed to 'version' to reduce width of column
name. The columns 'createdBy' and 'modifiedBy' are changed to user names
instead of profile IDs. The columns 'createdOn' and 'modifiedOn' are changed
to a date. Note that these columns do not need to exist in the fileview.
}
