% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-visualization.R
\name{data_summary}
\alias{data_summary}
\title{Get summary of data}
\usage{
data_summary(data)
}
\arguments{
\item{data}{Dataframe or tibble with file data.}
}
\value{
Tibble with summary information.
}
\description{
Get a summary of the data using the skimr
package, along with a custom function that appends an
extra column with a string showing each value for
a given variable and the number of times the value
appears. Returns a tibble with the data.
}
