% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-edit-annotations.R
\name{edit_annotations_server}
\alias{edit_annotations_server}
\title{Edit annotations module server}
\usage{
edit_annotations_server(id, fileview, annots_folder, syn, synapseclient)
}
\arguments{
\item{fileview}{The fileview for a specific study.}

\item{annots_folder}{Synapse folder ID to store generated annotation csvs in.}

\item{syn}{Synapse client object.}

\item{synapseclient}{Synapse client.}
}
\description{
Server function for the edit annotations module.
}
