% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fix_list}
\alias{fix_list}
\title{Fix character list}
\usage{
fix_list(data)
}
\arguments{
\item{data}{A list to fix.}
}
\value{
The fixed character vector.
}
\description{
Fix a list that should be all characters and
  should have `NA` fields, not `NaN` fields.
}
\keyword{internal}
