% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file-info.R
\name{format_date}
\alias{format_date}
\title{Format date}
\usage{
format_date(date)
}
\arguments{
\item{date}{Date in...}
}
\value{
Date as a string.
}
\description{
Format a date to be readable and print nicely.
}
