% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-all-file-data.R
\name{get_all_file_data}
\alias{get_all_file_data}
\title{Get data for all files in table}
\usage{
get_all_file_data(fileview, syn)
}
\arguments{
\item{fileview}{Tibble with columns 'id' and 'metadataType',
at minimum, and one row per file. 'id' is the synId
for the file.}

\item{syn}{Synapse client object.}
}
\value{
Original tibble with extra column 'file_data' of
  tibbles, holding the data from each file.
}
\description{
Downloads all the file data and attaches
to table as a tibble column.
}
