% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-all-file-data.R
\name{get_all_file_templates}
\alias{get_all_file_templates}
\title{Get the templates for all files in table}
\usage{
get_all_file_templates(fileview)
}
\arguments{
\item{fileview}{Tibble with columns 'metadataType', 'species',
and 'assay', at minimum, and one row per file.}
}
\description{
Gathers all the template synIds from the config file for each file
and attaches to table as a new column.
}
