% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_all_studies_table}
\alias{get_all_studies_table}
\title{Get synapse fileview}
\usage{
get_all_studies_table(fileview_id)
}
\arguments{
\item{fileview_id}{The synId for the full metadata Fileview.}
}
\value{
Tibble with the Fileview information.
}
\description{
Download a synapse fileview and return as a tibble.
}
