% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-file-indices.R
\name{get_file_indices_named}
\alias{get_file_indices_named}
\title{Get named file indices by metadataType}
\usage{
get_file_indices_named(study_view, meta_types)
}
\arguments{
\item{study_view}{The file view for a study, with the
columns 'file_data', where each element is a tibble
with the data, and metadataType, where each element
is the metadataType for the file.}

\item{meta_types}{List of metadataTypes.}
}
\value{
Named list of indices from `study_view` for which
  the metadataType exists; `NULL` if none exist.
}
\description{
Get the file indices for `study_view` based on metadataType
and return as list with file metadataType as name of the
elements.
}
