% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-study-overview.R
\name{get_info_box_ui}
\alias{get_info_box_ui}
\title{Get info box UI}
\usage{
get_info_box_ui(num_docs, num_meta, percent_success)
}
\arguments{
\item{num_docs}{Number of documentation files.}

\item{num_meta}{Number of metadata files.}

\item{percent_success}{Integer value between 0 and 100 representing the
success rate for validation.}
}
\description{
Creates the html for the info box UI.
}
