% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_most_recent_time}
\alias{get_most_recent_time}
\title{Get most recent time}
\usage{
get_most_recent_time(times)
}
\arguments{
\item{times}{Vector of POSIX times.}
}
\value{
The most recent time in `times`.
}
\description{
Get the most recent time from within
a list of POSIX times.
}
