% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_template}
\alias{get_template}
\title{Gets the template synId}
\usage{
get_template(metadata_type, species = NA, assay_type = NA)
}
\arguments{
\item{metadata_type}{The metadata type, or NA for manifest.}

\item{species}{The species.}

\item{assay_type}{The type of assay.}
}
\value{
synId for the template in Synapse.
}
\description{
Gets the template synId from the config file.
}
