% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file-info.R
\name{get_user_name}
\alias{get_user_name}
\title{Get the user name}
\usage{
get_user_name(user_id)
}
\arguments{
\item{user_id}{Synapse id as a string or number.}
}
\value{
User name as a string.
}
\description{
Get the user name based on Synapse Id.
}
