% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{main}
\alias{main}
\title{Gathering data and validating metadata files}
\usage{
main()
}
\description{
This function is here as an example workflow, for now.
It shows how the functions in this package work together
to get the information needed to monitor data curation
status for studies.
}
\details{
This function will most likely be adjusted, improved,
and eventually refactored during the switch to being
part of a shiny app.

Currently assumes user is registered with Synapse, and
has access to the fileview and files needed.
}
