% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-num-files-ids.R
\name{num_doc_files}
\alias{num_doc_files}
\title{Get the number of documentation files}
\usage{
num_doc_files(study_view)
}
\arguments{
\item{study_view}{The file view for the study files.
At minimum, the study_view is a dataframe or tibble
with `metadataType` as a column name. Any `NA` in
this column is assumed to be documentation files,
all others are assumed to be metadata files.}
}
\value{
The number of documentation files present in
  the `study_view`.
}
\description{
Get the number of documentation files
in a study.
}
