% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-num-files-ids.R
\name{num_manifest_files}
\alias{num_manifest_files}
\title{Get number of files in manifest}
\usage{
num_manifest_files(study_view)
}
\arguments{
\item{study_view}{The file view for a study, with the
columns 'file_data', where each element is a tibble
with the data, and metadataType, where each element
is the metadataType for the file.}
}
\value{
The number of unique files in manifest or
  0 if the  manifest is not in the `study_view`.
}
\description{
Get the number of unique files in manifest.
}
