% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{num_meta_files}
\alias{num_meta_files}
\title{Get the number of metadata files}
\usage{
num_meta_files(study_view, num_docs = NULL)
}
\arguments{
\item{study_view}{The file view for the study files.
At minimum, the study_view is a dataframe or tibble
with `metadataType` as a column name. Any `NA` in
this column is assumed to be documentation files,
all others are assumed to be metadata files.}

\item{num_docs}{Number of document files that are in
the `study_view`. If `NULL`, will call
`num_doc_files()` to get the value. This is here in case
that the number of document files is known already.}
}
\value{
The number of metadata files present in the `study_view`.
}
\description{
Get the number of metadata files in a study.
}
