% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-num-ids.R
\name{num_specimens}
\alias{num_specimens}
\title{Get number of specimens}
\usage{
num_specimens(study_view)
}
\arguments{
\item{study_view}{The file view for a study, with the
columns 'file_data', where each element is a tibble
with the data, and metadataType, where each element
is the metadataType for the file.}
}
\value{
Number of unique specimen ids.
}
\description{
Get the number of unique specimens for a study
from the metadata files.
}
