% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-results.R
\name{percent_pass_validation}
\alias{percent_pass_validation}
\title{Percent of passing validation results}
\usage{
percent_pass_validation(results)
}
\arguments{
\item{results}{A list of `dccvalidator` results,
which are of the type `check_pass`, `check_fail`,
or `check_warn`.}
}
\value{
Percent of validation checks passed, rounded
  to the nearest whole number. Returns 0 if `results` is `NULL` or if
  all checks in `results` are `NULL`.
}
\description{
Gives the overall percentage of validation
results that are passing.
}
