% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-server.R
\name{set_up_ui}
\alias{set_up_ui}
\title{Setup UI for all studies}
\usage{
set_up_ui(fileview)
}
\arguments{
\item{fileview}{A data frame with at least
one column, `study`.}
}
\description{
Setup UI for all studies in the fileview.
}
