% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-study-overview.R
\name{study_overview_ui}
\alias{study_overview_ui}
\title{UI for the study overview module}
\usage{
study_overview_ui(id)
}
\arguments{
\item{id}{Id for the module}
}
\description{
Creates the UI for the study overview
module.
}
