% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-visualization.R
\name{summarize_values}
\alias{summarize_values}
\title{Summarize values present}
\usage{
summarize_values(values)
}
\arguments{
\item{values}{The values to summarize in a list.}
}
\value{
String with the form "value1 (2), value2 (4)",
  where the value is given with the number of
  occurrences in parenthesis.
}
\description{
Get a list of values present and the number of times
each variable appeared.
}
