% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate-by-study.R
\name{validate_all_studies}
\alias{validate_all_studies}
\title{Validate all studies}
\usage{
validate_all_studies(fileview, annotations, syn)
}
\arguments{
\item{fileview}{The fileview table.}

\item{annotations}{A data frame of annotation definitions.
Must contain at least three columns: key, value, and columnType.}

\item{syn}{Synapse client object.}
}
\value{
Named list of check results, where the names of the list refer
  to the study. Returns NULL if no studies in fileview.
}
\description{
Validate metadata and manifest files for every study
in consortium. Returns a named list of validation results, where the name
refers to the study name. The validation checks are done via
`dccvalidator::check_all()`.
}
