% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate-by-file.R
\name{validate_biospecimen_meta}
\alias{validate_biospecimen_meta}
\title{Validation of biospecimen metadata}
\usage{
validate_biospecimen_meta(biospecimen, template, annotations)
}
\arguments{
\item{biospecimen}{Biospecimen metadata in dataframe.}

\item{template}{The biospecimen template.}

\item{annotations}{A data frame of annotation definitions.
Must contain at least three columns: key, value, and columnType.}
}
\value{
List of condition objects indicating whether the manifest
  has passed each validation check.
}
\description{
Validates the biospecimen metadata according to
specific requirements and returns list of results.
}
