% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate-by-file.R
\name{validate_individual_meta}
\alias{validate_individual_meta}
\title{Validation of individual metadata}
\usage{
validate_individual_meta(individual, template, annotations)
}
\arguments{
\item{individual}{Individual metadata in dataframe.}

\item{template}{The individual template.}

\item{annotations}{A data frame of annotation definitions.
Must contain at least three columns: key, value, and columnType.}
}
\value{
List of condition objects indicating whether the manifest
  has passed each validation check.
}
\description{
Validates the individual metadata according to
specific requirements and returns list of results.
}
