% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate-by-file.R
\name{validate_manifest}
\alias{validate_manifest}
\title{Validation of manifest}
\usage{
validate_manifest(manifest, template, annotations, syn)
}
\arguments{
\item{manifest}{Manifest data in dataframe.}

\item{template}{The assay template.}

\item{annotations}{A data frame of annotation definitions.
Must contain at least three columns: key, value, and columnType.}

\item{syn}{Synapse client object.}
}
\value{
List of condition objects indicating whether the manifest
  has passed each validation check.
}
\description{
Validates the manifest according to specific
requirements and returns list of results.
}
