% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate-by-study.R
\name{validate_study}
\alias{validate_study}
\title{Validate a study}
\usage{
validate_study(study_table, annotations, syn)
}
\arguments{
\item{study_table}{Tibble with fileview information for
a single study. Expected columns are: 'metadataType',
'file_data' (tibble column with data for each file), 'assay',
'species'. Should have, at most, one row per metadataType.}

\item{annotations}{A data frame of annotation definitions.
Must contain at least three columns: 'key', 'value', and 'columnType'.}

\item{syn}{Synapse client object.}
}
\value{
Named list of check results. Returns `NULL` if there are no rows.
}
\description{
Run validation checks for a single study and return the
results. Validation checks are run via `dccvalidator::check_all()`. It
is assumed that there is, at most, one row per metadataType in the
`study_table` passed in.
}
