% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-visualization.R
\name{visualize_data_types}
\alias{visualize_data_types}
\title{Visualize data types}
\usage{
visualize_data_types(data)
}
\arguments{
\item{data}{Dataframe or tibble with file data.}
}
\description{
Visualize the data class types, including
missing data, using the visdat and ggplot
packages.
}
