% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file-info.R
\name{create_info_table}
\alias{create_info_table}
\title{Create a nice, printable version of the file information}
\usage{
create_info_table(fileview)
}
\arguments{
\item{fileview}{The fileview table. Assumes the following
columns exist:
  name, id, currentVersion, metadataType, species,
  assay, createdBy, createdOn, modifiedOn}
}
\description{
Uses the fileview to create a table of relevant
file information.
}
