% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-edit-annotations.R
\name{edit_annotations_server}
\alias{edit_annotations_server}
\title{Edit annotations module server}
\usage{
edit_annotations_server(input, output, session, fileview)
}
\arguments{
\item{input}{Shiny input}

\item{output}{Shiny output}

\item{session}{Shiny session}

\item{fileview}{The fileview for a specific study.}
}
\description{
Server function for the edit annotations module.
}
