% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-edit-annotations.R
\name{edit_annotations_ui}
\alias{edit_annotations_ui}
\title{Edit annotations module UI}
\usage{
edit_annotations_ui(id)
}
\arguments{
\item{id}{Module id.}
}
\description{
UI function for the edit annotations module.
}
