% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study-table.R
\name{filter_study_table_latest}
\alias{filter_study_table_latest}
\title{Get study table with most recent metadata files}
\usage{
filter_study_table_latest(fileview, study_name = NULL)
}
\arguments{
\item{fileview}{The fileview.}

\item{study_name}{Name of study. If NULL, does not filter by study to
return study specific table.}
}
\description{
Filter fileview to get table with most recent metadata files.
The table will have all documentation files, but will only keep the most
recent version of files for the following metadata types: biospecimen,
manifest, assay, individual. Assumes that the fileview is for a single
study, but ultimately will keep most recent file information for each
metadata type, without filtering by study. If want to also filter by a
study name, specify name in `study_name` parameter.
}
