% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-results.R
\name{percent_pass_validation}
\alias{percent_pass_validation}
\title{Percent of passing validation results}
\usage{
percent_pass_validation(results)
}
\value{
Percent of validation checks passed, rounded
  to the nearest whole number.
}
\description{
Gives the overall percentage of validation
results that are passing.
}
