% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-study-overview.R
\name{study_overview_server}
\alias{study_overview_server}
\title{Server for the study overview module}
\usage{
study_overview_server(input, output, session, fileview, annotations)
}
\arguments{
\item{input}{Shiny input}

\item{output}{Shiny output}

\item{session}{Shiny session}

\item{fileview}{The fileview for a specific study.}

\item{annotations}{A dataframe of annotation definitions.}
}
\description{
Server for the study overview module.
}
