% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate-by-study.R
\name{validate_all_studies}
\alias{validate_all_studies}
\title{Validate all studies}
\usage{
validate_all_studies(fileview, annotations, syn)
}
\arguments{
\item{fileview}{The fileview table.}

\item{annotations}{A data frame of annotation definitions.
Must contain at least three columns: key, value, and columnType.}

\item{syn}{Synapse client object.}
}
\value{
The fileview table with a new column, 'results', with
  a list of validation results for every file.
}
\description{
Validate metadata and manifest files for every study
in consortium. Returns the fileview table with a new 'results'
column that contains a list of validation results for each
file.
}
