% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate-by-study.R
\name{validate_study}
\alias{validate_study}
\title{Validate a study}
\usage{
validate_study(study_table, annotations)
}
\arguments{
\item{study_table}{Tibble with fileview information for
a single study. Expected columns are: 'metadataType',
'file_data' (tibble column with data for each file), 'assay',
'species', and 'template'.}

\item{annotations}{A data frame of annotation definitions.
Must contain at least three columns: 'key', 'value', and 'columnType'.}
}
\value{
study_table with a new column, 'results', that holds the
  result lists for each file in the table.
}
\description{
Run validation checks for a single study and appends the results
to the study_table in a new 'results' column. The results
are in list form with specific names for the tests.
}
